Move rocketchat-notifications to a specific package
