+++ .meteor/packages
@@ -205,3 +205,4 @@ rocketchat:settings
 rocketchat:models
 rocketchat:metrics
 rocketchat:callbacks
+rocketchat:notifications
