+++ packages/rocketchat-lib/package.js
@@ -29,6 +29,7 @@ Package.onUse(function(api) {
 	api.use('rocketchat:migrations');
 	api.use('rocketchat:metrics');
 	api.use('rocketchat:callbacks');
+	api.use('rocketchat:notifications');
 	api.use('rocketchat:accounts');
 	api.use('modules');
 	api.use('rocketchat:i18n');
