+++ packages/rocketchat-notifications/client/index.js
@@ -1,0 +1,5 @@
+import Notifications from './lib/Notifications';
+
+export {
+	Notifications,
+};
