+++ packages/rocketchat-notifications/server/index.js
@@ -1,0 +1,5 @@
+import Notifications from './lib/Notifications';
+
+export {
+	Notifications,
+};
